Remove unncessary abstraction over the soci index builder module
Improvements should be made to that module instead.
